backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic>=0.45.0
python-dotenv==1.0.1
pydantic[email]>=2.9.0
edge-tts>=6.1.0
httpx>=0.27.0
# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
# Redis
redis>=5.0
# Auth
jinja2>=3.1
python-jose[cryptography]>=3.3
bcrypt>=4.1
python-multipart>=0.0.9
# Calendar (optional)
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
